チーム作成のViewを分割
